Changed default initial setting for Genre codes
Since the IPTC Genre codes are human-readable English words that may
already exist in the Keywords for some images, I’ve turned off that we
scan the Keywords for Genre codes by default. This means that if we
have a photo with a few Genre codes carefully entered, it won’t be
overwritten by a single, possibly inaccurate term that happens to
overlap with the Genre codes (e.g. ‘Background’). In the LoweMo vocab,
this issue is mitigated by the fact that most keywords are lower case
(other than the Genre codes and some other words that are properly
capitalized in the English language)