Experimental branch, plus some fixes
Most of the new stuff in this branch will be rolled into a separate
script or plugin, but was added here for convenience’s sake
(functionality to identify redundant/duplicate keywords, i.e. keywords
that share the same name within the keyword tree). Other fixes have
also been made in this branch, though, including removal of trailing
whitespace, removal of commented out code, addition of error messages
to be shown if configured parent terms (for Genre terms) do not exist,
so that we do not try to get children for non-existent terms and with
an error message that a user might not understand, etc.