Misc fixes. Got new features working.
The major new features 1) to provide a setting for automatically adding
parent keywords for any selected keywords (not just on export), 2) to
provide a setting for adding any direct child code keywords, e.g. if
‘architecture’ is selected, add corresponding IPTC code, ‘20000032’,
and 3) refactoring to share LUTILS and KwUtils — this may be further
refactored to allow that to be a separate module. Still need more
documentation updates for new features, but code is working.