Currently working, but will further refactor...
This is a non-broken state that I can roll back to if things get into
worse shape (or to diff against, if needed). I have determined that
it’s not possible to have two progress scopes running within the same
withWriteAccessDo process and have decided that it makes more sense to
combine both loops, for the sake of being able to undo (if multiple
withWriteAccessDo processes are completed without user input, they are
bundled into an irreversible catalog change, which is not ideal. So in
one loop, I will do both the adding of keywords and the copying of them
to the IPTC fields. Since the catalog is not updated until the end, we
need to track which codes are added for the purpose of copying
newly-added codes to a photo’s IPTC fields.