Refactored to put all processes in one progress scope
Multiple progress scopes in the same catalog:withWriteAcccessDo block
does not work well, so now we do one for loop with all processing.
Everything works nicely now, except that the progress captions (for
sub-tasks) do not get displayed. No errors, though, and this is a
relative non-issue since we do at least have a visible and cancelable
process.